[vim] Resolve double-loading mystery.
It so happens that aniseed loads the init module in the end of env
setup. If no module is specified, this will result in loading `init.lua`
twice, once before and once after aniseed is loaded. Before I needed to
avoid loading fennel-based plugins on the first run because `package.path`
is not set yet at that stage.
A better solution is to move loading of fennel-based plugins into a
separate module `startup` and ask aniseed to load it explicitly.